New fading features.
createGraphicsAssets now creates RGBA versions of the fadeMasks.
Debug now has buttons to select a fade.
tilemapToImageData now does tile fading on-the-fly. Handles colorData prior to the fade. (fadeImageData)
Added updateBG1BgColorRgba which is specific to the bgColor of the BG1 layer.
The LayerObject class now has a hidden property that will hide the drawing but leave it's data in the instance.
Can draw game cards, draw pile, discard pile.