Refactoring, work on debug.
_DEBUG:
Update the CSS for the debug tables.
Added timings display for Loop, Logic, Draw.
Added display for layers that had a change.

uno.js:
Added the missing await for the call to _APP.game.gameLoop.init().

uno_main.js:
Refactor of the layout of the code for easier access to the mostly likely to change parts.
Added much more commenting to explain the gameLoop variables.
The parts of the loop are timed: Loop, Logic, Draw.
Debug is now called from endOfLoopTasks.

gfx.js:
Added _GFX.layerObjs.createPlaceholder. It will create an empty object for a future layer object.