JSGAME fixes, moving of base classes to gfx.js, update error handler.
Moved body CSS to #wrapper. Works better with JSGAME.
Removed _GFX.layerObjs.createPlaceholder. Not needed.
Removed support for the hidden flag in LayerObjects. Never used.
Moved the LayerObject and PrintText classes to gfx.js (_GFX).
 * These are the LayerObject classes that will be considered "built-in" to the graphics engine..
Commenting added to _WEBW_V.
Moved the error_display to be inside #wrapper so that JSGAME would be able to see it.
Updated the game init functions (JSGAME/StandAlone) to load the error handler after the HTML is loaded.
_APP.debugActive is now set via URL params.
Gamestates are now changed in the endOfLoopTasks function.