Improvement to graphics drawing, fixed hashCacheMap management.
Added a clearRegion function.
CLEAR.oneLayerData now properly removes mapKeys.
CLEAR.oneMapKey removes from the hashCache when the removeHashOnRemoval flag is set. Additionally, it will remove related hashes (hashCacheHash_BASE) which is a hash that does not include the settings data, thus, "BASE".
The fadeBeforeDraw flag is set only if a global fade is active. It indicates that new/changed drawings need a fade applied before they are drawn.
New timestamps with "timeIt".
Fades are primarily applied in drawImgDataCacheFromDataCache and not applied to the whole output as before.
New checks in the _GFX.funcs.layerObjs functions. Functions immediately return if the gamestate is "".
removeOneAllGamestates will remove a specified key in objs for all gamestates. Unused currently.
Fixed the progress bars. Removed TOTAL and replaced it with DEBUG.
Completed the DEBUG for deckControl.