Hashing and rotation improvements/fixes.
Improvements to the Card class.
Creation of the Deck class.
Creation of the Gameboard class.
If the gamestate is set to "" then the loop will be skipped.
errorHandler changes.
hashCacheHash and hashCacheHash_BASE are generated by the WebWorker only.
Improvements to createImageDataFromTilemap.
Fixed the distorted rotation.
Fixed the bug where rotations could result in clipping of the image.