The Layer Object Editor can now display additional features depending on the className of the LayerObject.
All classes now save their className to this.className in the class.
 * This might not actually be needed.
Improvements to the Card class to keep the state of a card correct when switching colors/values.
Updates to the Deck class.
The Layer Object Editor can now display additional features depending on the className of the LayerObject.
 * The Card class offers the ability to change the card size, color, and value.