Added origin keys for hashCache, fixed gamestate change prevention bug.
Added origin keys for the hashCache. It can now be seen where a tilemap image came from and if it is a modification of an existing tilemap image.
Corrected the default setting for removeHashOnRemoval in some places.
Fixed bug where gamestates could not be changed when not in debug mode.
Removed some maps from the main thread generateCoreImageDataAssets since they will be generated by the INIT_PRECACHE anyway.
Updates to updateProgressBar2.