Refactor of DEBUG.
Correction to the origin values for custom maps.
differedProms now actually resolve with their data (this was not necessary until the new DEBUG.)
Changed how debug mode is allowed. (Needs to be localhost now.)
DEBUG runs after the game loop and requests data instead of just receiving it when requesting a draw.
The text value of text tilemaps is sent now which is helpful for debug.
DEBUG progress bars are no longer CSS/JavaScript. They are now Canvas/JavaScript.