Improvements to the DEBUG performance.
Improvements to the new canvas-based progress bars.
 * Range separators.
 * Colored rectangle is colored to match the range.
Additional data checks in the hashCache debug.
Additional data checks in the drawTimings debug.
If you are on a debug tab that does not require updates then debug will be shorter.
 * Default now is the Fades tab.
awaitDraw is set true when debug mode is active.
Added a "set" func to timeIt.
When requesting _DEBUG.updateDebugTimings the returned ALLTIMINGS be be reduced on each call
 * So that values do not get stuck when there have not been any updates.