Fixed GameBoard acceptCursor, handling of card rows during the initial deal.
New debug Player/Row display shows the current row for each player.
Addition of countColorMatches for a player's hand.
_resetCardPositions function to make sure that the LayerObjects that represent the displayed cards can be reset.
Update to ColorChanger show that will determine the dominant color in the player's hand and default the cursor to that color.
The currentRow is now stored per player instead of globally.
When using setNextPlayer the resetting of the currentRow value is now optional via function argument.
Fixed bug in GameBoard acceptCursor that would return a copy of the LayerObject but not the complete object.
In gs_PLAYING.js moved flags/values to the top of the object.
When performing the initial card deal when the initDealPos reaches 5 the row is changed and this is visually reflected on the screen.
 * It was just placing cards over existing cards before. It is cleaner now.
Cards now have a validity check before selection.