Fixes to Deck and Card functions. Can now change the displayed row.
Debug2 flags now displays last played and last drawn cards.
Fixed bug in debug2 that was not checking for a previous value match before updating the DOM.
Fixed bug in the Card class that was not setting this.changed after a change.
 * _change_value, _change_color, _change_size, change_wholeCard.
 * This only became apparent when switching rows and the new card with the same value in the same position would not be recolored.
  * This is because it was not detected as a change. Since the value did not change neither did the tilemap.
  * Fixed.
Fixed bug in flipPlayerCardsUp, flipPlayerCardsDown where it would not slice the correct cards based on the row.
Can now pass a turn and cancel a card selection.
Can now change the displayed row on the player's turn.